Don't print ref (can't parse it), don't use ASNode yet
Printing refs doesn't really work, since there is no equivalent on the
other side for parsing them in an environment (I guess it could be
added, but if we're not printing refs on object fields, why not be
consistent).

Change the ASNode stuff I put in to just be HTMLElements for now, for
consistency across what exists.