quest() types with cheats.
Need variable arity to handle Array.slice()

Need to handle the lookups on hunter and pecker.

This is actually annoying because they are dictionary types, and don't
necessarily have the "right" types.  Think about it this way ---
nothing proves (yet) that in

hunter[selector.op]

selector.op !== "valueOf"

So we can't say yet that hunter (even as a dictionary type) will for
sure give back the function type we expect.  If we use an ObjStar for
it, it will say that toString is one of the possible things we could
return, and then applying the hunter function won't work (and
shouldn't in the TypedJS sense --- even though it might be safe from
the security perspective).

More thoughts needed.