Move some ref things around.
ObjStar lookup was unsound --- objstar-update was passing.

I'm more sure that allowing un_ref to work over unions is sound, but
there could be something else that I'm missing.

This makes it so (ref (deref e)) does not have the same type as e, for
example if e : ref t1 + ref t2.  The larger expression will type to
ref t1 + t2.

I think this is just weird, though, not unsound.