Add ObjCast to the type checker.
The ObjStar type can be legitimately introduced only as a cast on ENew
and EObject.  These are the only places where *all* of the fields of
the object are known by the type system.  Just having an object-type
annotation leaves all the other fields unspecified (so they might as
well by Any), making subtyping in the Object -> ObjStar direction
hopeless.

ENew and EObject have an implicit upcast from ObjStar to Object with
the current typing rules --- an object literal is really an ObjStar
with all the named fields, Object as its prototype, and an Undef
*-type, but the extra information is dropped.  Similarly, in a new
expression, the protoype information is available and all other fields
are undefined.  Once duck typing gets involved and information about
fields is dropped, it is impossible to recover other than saying Any,
which isn't acceptable for the precision we want out of ObjStar.

So, we introduce a casting annotation that can be used at these
specific types of expression.  Things like:

var o = /* obj* {x:Int, #proto: Object, *: Undef, #code: Undef} */
        {x:5};

Are legitimate, even though the type {x:Int} isn't a subtype of the
ObjStar type.

In terms of ADsafe, what we want is to use the obj* cast on objects
that are going to reach the client (like new Bunch(), and event =
{...}), and let ADsafe use objects as usual internally.