Several changes to the type checker (read whole message)
1.

Missed field lookups on TConstrs that correspond to primitives are
ignored.  I changed the operation to type to TBot instead of TRef
(Undef), because they don't actually have anything meaningful coming
out.  If you do:

a = "asdf";
a.foo = 5;
a.foo; // foo is undefined

So technically these go to undefined, but that was screwing up typing
things like

o.foo = bar;

Where o is the Ad type, because if o is a string and o.foo produces
TRef (Undef) as part of the union, this would force bar to be Undef to
type, even though the assignment has no effect on the value of the Str
part of the Ad union.  There might be a better solution for this, but
what I've done is obvious and blatant to change if it's wrong (the
missed-field clause of bracket for TConstr).

2.

Applying inapplicable values.  I added a function applicable and a
function to filter out inapplicable values.  Right now, I just say
that no TConstr type is applicable.  This is for applying the Ad type
--- once again Str, Undef, etc. are in the union, and always cause
runtime errors when applied.  So, it's safe to take them out of a
union that is in the function application position, and check the
value that remains (which will be the part of the Ad type we care
about).

3.

The Obj* cast.  The rules for casting an object literal or a new
expression to an obj* are:

  i.   Check that the #proto of the obj* and the prototype of the object
       are the same (for literals, just Object, for new, whatever the
       super-constructor of the constructor is).

  ii.  Get the list of fields from the object literal or the
       constructor-type itself, and for each check that either:

       a.  It is present in the obj* type and a subtype of the field in
          the obj* type (since they're references, this is basically
          equivalence)
       b.  It is a subtype of the *-type for the obj*

  iii. Ensure that all the other named fields on the obj* supertype
       Undef, because all the other fields on the literal or new
       object will be Undef.

This seems to be working ok for the_event in dom_event, and for
Bunches.  We'll see what else comes up.