Add support for *absent* fields.
An object can look like this:

{*: T, #proto: Tp, #code: Tc, x: _}

And the underscore means x *cannot* be an own property of the object.
If an object o has this type, the expression:

o.x

Will type to fields(Tp, ${"x"}).  It won't union with the star-type.

This is a more accurate representation of things like valueOf in
ADsafe, which actually enforces this --- objects in the ad don't have
valueOf fields, and all use the immutable ones on their prototypes.

Also, absent fields cannot be assigned to.