Try and print synonyms if you can.
Changed the pretty-printer for types to take a function argument that
transforms from typ->typ.  Then, added string_of_typ' to Env.Pretty,
which takes an environment and tries to find synonyms in the
environment that match up with the type before printing.

This gives way better error reporting and makes a lot of things more
legible.  Doesn't work tremendously well on the final environment, but
works for the 'Ad type in many cases so it's at least a little bit
nice.

I weakened the diff of environments to try and get more out of the
ending environment, not sure how that will go.