On TArrows, differentiate between functions that bind this or don't
Administrative lambdas don't bind this --- the analysis is currently
being conservative and treating them as TTop.  This is fine, but in
the type checker this is trouble for instances of this within nested
loops.  So we allow function types to specify this-ness or not,
depending on if they should bind this in their body.

Basically just an implementation optimization instead of desugaring to
explicit this arguments everywhere.