The last change I'll make to the type language
So some libary functions return new objects, even though they aren't
syntactic object literals, new expressions, or array literals.  But,
they're new to the program so they can be objcast.

So, we invent the type Fresh, so this can be expressed and used.  For
example:

fresh {x: ref Int} <: {x: ref Num}

Because the object is guaranteed to be new.  This is necessary for
typing RegExp.exec, because it returns a fresh array.