Break typed-adsafe's typing a little bit, but make sure that it works
The sudoku widget works with all the changes I've made to adsafe.js.
I had to break typing a little bit to make this happen, so I'd like to
get this version of typed-adsafe.js to type to be the thing I send to
Doug.

I had to make changes in the widgets because I made some mistakes
copying them somehow.  These are updated to reflect what's on the web,
and still pass typedjs_JSlint.