fix fields by adding overlap, make regexes subtypes of strings
Now TRegex <: Str, and string literals parse to singleton string
regexes.

Also, we were doing the wrong operation for fields.  We do not want
containment of one regex in another, we want emptiness of the
intersection.  If the regex in the brackets intersected with a field
regex is non-empty, we need to include that field.  If it is empty,
then that field can be ignored.