some fixes to regexps
Make it so complemented sets don't include the # character.  Then,
Leaf.partition can keep track of everything going out (including
possible # characters), and what goes to an error state.

Also, added find_word and counterexample, which are like nullable and
contains but provide a word that is in the intersected dfa