took a stab at arrays (longer comment below)
When we got rid of TConstr, we also lost the type arguments that came
with it.  This wasn't necessary for the environment or typing array
literals (an ETypApp would suffice), but its a little trickier to
write function signatures with arrays like:

Array<Int> -> Array<Int>

So I added this TApp thing, that takes two types and during check_typ
will substitute the second into the first if the first is a forall.
It's just an annotation-level place to enforce a type substitution,
but I'm not sure how wonky it is.  Added some array test cases, though
the old ones still fail since I can't look up integers yet.