Inherited fields work in one direction
Also fixes to subtype_object and fields:

Subtype_object had a TODO: check overlap --- this is now fixed and
object subtyping won't allow the RHS to have more strings than the LHS
(except for inherited fields)

Fields had a similar problem --- we were going into proto when we
shouldn't have (e.g. we allowed lookups of y on the type:

{__proto__: {y : Bool}}

which shouldn't be allowed).

Changed proto_str to __proto__

Moved fields inside Env, because it is used by subtype_object (not
sure if this will make proofs easier or harder).